fix and add tests for delta format translation
