+++ app/gui2/package.json
@@ -16,7 +16,7 @@
     "test": "vitest run",
     "build-only": "vite build",
     "compile-server": "tsc -p tsconfig.server.json",
-    "typecheck": "vue-tsc --noEmit -p tsconfig.vitest.json --composite false",
+    "typecheck": "vue-tsc --noEmit -p tsconfig.app.json --composite false",
     "lint": "eslint .",
     "format": "prettier --write src/ && eslint . --fix",
     "build-rust-ffi": "cd rust-ffi && wasm-pack build --release --target web",
@@ -24,6 +24,7 @@
   },
   "dependencies": {
     "@babel/parser": "^7.22.16",
+    "@fast-check/vitest": "^0.0.8",
     "@open-rpc/client-js": "^1.8.1",
     "@vueuse/core": "^10.4.1",
     "codemirror": "^6.0.1",
