+++ app/gui2/tsconfig.server.json
@@ -1,11 +1,12 @@
 {
   "extends": "@tsconfig/node18/tsconfig.json",
   "include": ["ydoc-server/**/*", "shared/**/*"],
+  "exclude": ["ydoc-server/**/__tests__/*", "shared/**/__tests__/*"],
   "compilerOptions": {
     "module": "ESNext",
     "moduleResolution": "Bundler",
-    "outDir": "../../node_modules/.cache/tsc",
     "composite": true,
+    "outDir": "../../node_modules/.cache/tsc",
     "types": ["node", "vitest/importMeta"]
   }
 }
